Rewrite to use rosbags package
