+++ Dockerfile.ros2
@@ -1,15 +1,19 @@
 FROM ros:humble-perception
 
-RUN apt-get update && apt-get install -y --no-install-recommends \
-    ffmpeg && \
-    rm -rf /var/lib/apt/lists/*
+# Install system packages
+RUN apt-get update \
+    && apt-get install -y --no-install-recommends \
+        ffmpeg \
+        python3-pip \
+    && pip3 install rosbags
+    && rm -rf /var/lib/apt/lists/*
 
 # Setting up working directory 
-WORKDIR /ros2bag2video_workspace
+WORKDIR /rosbag2video_workspace
 
 # Add user
 RUN adduser --quiet --disabled-password user
-RUN chown -R user:user /ros2bag2video_workspace
+RUN chown -R user:user /rosbag2video_workspace
 USER user
 
 ENTRYPOINT ["/ros_entrypoint.sh"]
